Add TSX highlighting tests
Add plenary tests to verify that TSX is highlighted in rainbow colors
correctly.

Uses a DSL for the source code to be highlighted:

* Each line that has at least 1 highlighted character needs a line below
  that specifies the rainbow level of that highlight.
* The line with rainbow levels needs to have `hl` and then numbers or
  spaces. This line will most likely be a comment line to avoid
  treesitter parser errors.
* The numbers must appear right below the character that should be
  highlighted (in the same column).

This makes writing test cases easy since one does not have to worry
about column and line numbers. The tests are very visual.

The tests are not complete yet. They do not robustly test the highlights
(extraneous highlights can slip through). Moreover, they will be
generalized to make defining new test cases as easy as defining a new
file with the highlight specification.