Detect extraneous and surrounding extmarks in the tests
Refactor the tests to detect extraneous and surrounding extmarks.

Extraneous extmarks are highlighted symbols that are not expected in the
test source code DSL.

Surrounding extmarks are extmarks that surround a given symbol that is
expected to be highlighted. This is necessary because
nvim_buf_get_extmarks returns only extmarks that are defined within a
given range. If an extmark wraps a given range, it is not returned. This
caused some TSX highlight tests to fail because `div` and other JSX
elements were highlighted with a single extmark and asking for an
extmark for the range of the `i` character (the middle one) returned 0
extmarks, since the extmark that highlights that range surrounds the
`i`.

This required refactoring the tests to retrieve all buffer extmarks and
check if all of them match the expected highlighted symbols and if there
are no more leftover extmarks after matching them up with the
expected ones.